build: fix for modern systems
Add a couple of missing header files, and document a workaround
for missing prototypes for `yylex`.

Versions of C >C99 really want `yylex` to have a proper function
prototype.  Unfortunately, `bison` does not generate one
automatically; furthermmore, the actual signature can change
depending on how it is invoked (e.g., whether using a reentrant
parser or not).

We don't much care, but POSIX says that giving the `-y` command
line option _should_ cause such prototypes to be generated.  But
the Bison maintainers ran into problems with that, so they
disable generating said prototypes unless the `POSIXLY_CORRECT`
environment variable is set.

Also, the Flex library is not necessarily `-ll`; on Linux, it
may well be `-lfl`.

Document this in the Makefile; kinda.

Signed-off-by: Dan Cross <cross@gajendra.net>